Update README.md
The data is collected from hand-written church records in Sweden. The church was responsible for all population registration until around 1870. Only death records have been analyzed. Since Cholera has around 40% CFR we can assume a larger number of affected but surviving individuals. This dataset illustrates the speed of a cholera pandemic in the 19th century in an urban environment with very rudimentary sanitation, and related traumatic effects on society and individuals.